+++ extras/tests/JsonVariant/or.cpp
@@ -7,7 +7,7 @@
 
 TEST_CASE("JsonVariant::operator|()") {
   DynamicJsonDocument doc(4096);
-  JsonVariant variant = doc.to<JsonVariant>();
+  JsonVariant variant = doc["value"].to<JsonVariant>();
 
   SECTION("undefined") {
     SECTION("undefined | const char*") {
@@ -23,6 +23,27 @@ TEST_CASE("JsonVariant::operator|()") {
     SECTION("undefined | bool") {
       bool result = variant | true;
       REQUIRE(result == true);
+    }
+
+    SECTION("undefined | ElementProxy") {
+      doc["array"][0] = 42;
+
+      JsonVariantConst result = variant | doc["array"][0];
+      REQUIRE(result == 42);
+    }
+
+    SECTION("undefined | MemberProxy") {
+      doc["other"] = 42;
+
+      JsonVariantConst result = variant | doc["other"];
+      REQUIRE(result == 42);
+    }
+
+    SECTION("ElementProxy | ElementProxy") {
+      doc["array"][0] = 42;
+
+      JsonVariantConst result = doc["array"][1] | doc["array"][0];
+      REQUIRE(result == 42);
     }
   }
 
@@ -43,6 +64,20 @@ TEST_CASE("JsonVariant::operator|()") {
       bool result = variant | true;
       REQUIRE(result == true);
     }
+
+    SECTION("null | ElementProxy") {
+      doc["array"][0] = 42;
+
+      JsonVariantConst result = variant | doc["array"][0];
+      REQUIRE(result == 42);
+    }
+
+    SECTION("null | MemberProxy") {
+      doc["other"] = 42;
+
+      JsonVariantConst result = variant | doc["other"];
+      REQUIRE(result == 42);
+    }
   }
 
   SECTION("int | const char*") {
@@ -54,6 +89,20 @@ TEST_CASE("JsonVariant::operator|()") {
   SECTION("int | uint8_t (out of range)") {
     variant.set(666);
     uint8_t result = variant | static_cast<uint8_t>(42);
+    REQUIRE(result == 42);
+  }
+
+  SECTION("int | ElementProxy") {
+    variant.set(42);
+    doc["array"][0] = 666;
+    JsonVariantConst result = variant | doc["array"][0];
+    REQUIRE(result == 42);
+  }
+
+  SECTION("int | MemberProxy") {
+    variant.set(42);
+    doc["other"] = 666;
+    JsonVariantConst result = variant | doc["other"];
     REQUIRE(result == 42);
   }